requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
torch==2.10.0+cu128
torchvision==0.25.0+cu128
Pillow==10.0.0
python-multipart==0.0.22
